LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-createAnswer-expected.txt


PASS createAnswer() with null remoteDescription should reject with InvalidStateError
PASS createAnswer() returns RTCSessionDescriptionInit
PASS createAnswer() when connection is closed should reject with InvalidStateError
